+++ database/migrations/2018_12_26_172524_add_payment_method_id_to_orders_table.php
@@ -14,7 +14,7 @@ class AddPaymentMethodIdToOrdersTable extends Migration
     public function up()
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->unsignedInteger('payment_method_id')->after('shipping_method_id')->index();
+            $table->unsignedBigInteger('payment_method_id')->after('shipping_method_id')->index();
 
             $table->foreign('payment_method_id')->references('id')->on('payment_methods');
         });
